chore(zip): Remove dependency on @loaders.gl/core
