+++ modules/loader-utils/src/lib/file-provider/file-handle-file.ts
@@ -1,4 +1,4 @@
-import {FileProvider} from '@loaders.gl/zip';
+import {FileProvider} from './file-provider';
 import {FileHandle} from './file-handle';
 
 /**
