+++ modules/loader-utils/src/types.ts
@@ -69,37 +69,3 @@ export type BatchableDataType =
   | Iterable<ArrayBuffer>
   | AsyncIterable<ArrayBuffer>
   | Promise<AsyncIterable<ArrayBuffer>>;
-
-/**
- * A FileSystem interface can encapsulate a FileList, a ZipFile, a GoogleDrive etc.
- */
-export interface IFileSystem {
-  /**
-   * Return a list of file names
-   * @param dirname directory name. file system root directory if omitted
-   */
-  readdir(dirname?: string, options?: {recursive?: boolean}): Promise<string[]>;
-
-  /**
-   * Gets information from a local file from the filesystem
-   * @param filename file name to stat
-   * @param options currently unused
-   * @throws if filename is not in local filesystem
-   */
-  stat(filename: string, options?: object): Promise<{size: number}>;
-
-  /**
-   * Fetches a local file from the filesystem (or a URL)
-   * @param filename
-   * @param options
-   */
-  fetch(filename: string, options?: object): Promise<Response>;
-}
-
-type ReadOptions = {buffer?: ArrayBuffer; offset?: number; length?: number; position?: number};
-export interface IRandomAccessReadFileSystem extends IFileSystem {
-  open(path: string, flags: string | number, mode?: any): Promise<any>;
-  close(fd: any): Promise<void>;
-  fstat(fd: any): Promise<object>;
-  read(fd: any, options?: ReadOptions): Promise<{bytesRead: number; buffer: Buffer}>;
-}
