+++ modules/zip/src/parse-zip/local-file-header.ts
@@ -1,4 +1,4 @@
-import {FileProvider} from '../file-provider/file-provider';
+import {FileProvider} from '@loaders.gl/loader-utils';
 
 /**
  * zip local file header info
